Se añade primer endpoint y mas cosas q no recuerdo en este punto estoy vibeando we vibe
