--- Videogames/MealtimeMayhem/Assets/Scenes/Scenes/Login.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 633b0b3c007013647be41af9c70a5c1c
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
