--- Videogames/MealtimeMayhem/Assets/Scenes/Scenes/SceneMau.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 945f38414be6d4140ad30313f5dbd968
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
