+++ Videogames/MealtimeMayhem/Assets/Scenes/Sign Up.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8d489246195dbc244b44ab0433ade33b
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
