--- Videogames/MealtimeMayhem/Assets/Scenes/TitleScreen.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 83edfd554a4a79b4e8e9608774146e05
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
