+++ Videogames/MealtimeMayhem/Assets/Scripts/API/UserCheck.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1e344307f72f86d498dc659259ea6515
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
