--- Videogames/MealtimeMayhem/Assets/Scripts/TestAPI/ButtonActions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dcfd50422d36c144a9d0175c142c0ab2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
