--- Videogames/MealtimeMayhem/Assets/Scripts/TestAPI/TestAPI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8cca8ccd410dd44f088de09dfdfc9f27
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
