--- Videogames/MealtimeMayhem/Assets/Scripts/TestAPI/VarTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a3e148d8a87ad4332a505488f606dfdc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
